+++ CHANGELOG.md
@@ -47,6 +47,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :bug: Bug Fix
 
 * Provided the `lang` attribute to `<html>` `iStaticPage`
+* Fixed a bug when providing `href` to `dataProvider` `bButton`
 
 ## v3.0.0-rc.210 (2021-07-07)